freebsd: Add param header

+++ src/platform/freebsd-types/sys/syslimits.rs
@@ -13,23 +13,23 @@ pub const ARG_MAX: usize = 2 * 256 * 1024;
 pub const ARG_MAX: usize = 256 * 1024;
 
 /// max simultaneous processes
-pub const CHILD_MAX: i32 = 40;
+pub const CHILD_MAX: usize = 40;
 
 /// max bytes in term canon input line
-pub const MAX_CANON: i32 = 255;
+pub const MAX_CANON: usize = 255;
 /// max bytes in terminal input
-pub const MAX_INPUT: i32 = 255;
+pub const MAX_INPUT: usize = 255;
 /// max bytes in a file name
-pub const NAME_MAX: i32 = 255;
+pub const NAME_MAX: usize = 255;
 /// max supplemental group id's
-pub const NGROUPS_MAX: i32 = 1023;
+pub const NGROUPS_MAX: usize = 1023;
 
 /// max open files per process
-pub const OPEN_MAX: i32 = 64;
+pub const OPEN_MAX: usize = 64;
 
 /// max bytes in pathname
-pub const PATH_MAX: i32 = 1024;
+pub const PATH_MAX: usize = 1024;
 /// max bytes for atomic pipe writes
-pub const PIPE_BUF: i32 = 512;
+pub const PIPE_BUF: usize = 512;
 /// max elements in i/o vector
 pub const IOV_MAX: usize = 1024;
